very big architecture update about getting price
git-svn-id: svn://192.168.0.55/repo@143 3dcf240c-a488-453e-83cb-a6019b294898

+++ voyanga/frontend/components/MFlightSearch.php
@@ -7,7 +7,7 @@
  */
 class MFlightSearch extends CComponent
 {
-    public static function getOptimalPrice($fromCityId, $toCityId, $date, $returnDate=false)
+    public static function getOptimalPrice($fromCityId, $toCityId, $date, $returnDate=false, $forceUpdate = false)
     {
         $flightSearchParams = new FlightSearchParams();
         $departureDate = date('d.m.Y', strtotime($date));
@@ -44,8 +44,11 @@ class MFlightSearch extends CComponent
             $criteria->addCondition('returnDate BETWEEN STR_TO_DATE("'.$returnDate.' 00:00:00", "%d.%m.%Y %H:%i:%s") and STR_TO_DATE("'.$returnDate.' 23:59:59", "%d.%m.%Y %H:%i:%s")');
             $criteria->addSearchCondition('withReturn', 1);
         }
-        $result = FlightCache::model()->deleteAll($criteria);
-        $fs->sendRequest($flightSearchParams);
+        if ($forceUpdate)
+        {
+            $result = FlightCache::model()->deleteAll($criteria);
+            $fs->sendRequest($flightSearchParams);
+        }
         $result = FlightCache::model()->find($criteria);
         if ($result)
         {
